Update huobi, gateio api endpoint

+++ huobi/src/main/java/com/fafabtc/huobi/data/remote/dto/HuobiMarketDetail.java
@@ -79,13 +79,13 @@ public class HuobiMarketDetail {
         @SerializedName("high")
         private double high;
         @SerializedName("id")
-        private int id;
+        private long id;
         @SerializedName("count")
-        private int count;
+        private long count;
         @SerializedName("low")
         private double low;
         @SerializedName("version")
-        private int version;
+        private long version;
         @SerializedName("vol")
         private double vol;
 
@@ -121,19 +121,19 @@ public class HuobiMarketDetail {
             this.high = high;
         }
 
-        public int getId() {
+        public long getId() {
             return id;
         }
 
-        public void setId(int id) {
+        public void setId(long id) {
             this.id = id;
         }
 
-        public int getCount() {
+        public long getCount() {
             return count;
         }
 
-        public void setCount(int count) {
+        public void setCount(long count) {
             this.count = count;
         }
 
@@ -145,11 +145,11 @@ public class HuobiMarketDetail {
             this.low = low;
         }
 
-        public int getVersion() {
+        public long getVersion() {
             return version;
         }
 
-        public void setVersion(int version) {
+        public void setVersion(long version) {
             this.version = version;
         }
